+++ Vault/src/utopia/vault/nosql/access/single/model/distinct/LatestModelAccess.scala
@@ -1,11 +1,8 @@
 package utopia.vault.nosql.access.single.model.distinct
 
-import utopia.flow.generic.model.immutable.Value
-import utopia.vault.nosql.access.single.model.SingleRowModelAccess
-import utopia.vault.nosql.access.template.model.DistinctReadModelAccess
-import utopia.vault.nosql.factory.row.FromRowFactoryWithTimestamps
-import utopia.vault.nosql.view.FilterableView
-import utopia.vault.sql.Condition
+import utopia.vault.nosql.factory.row.FromTimelineRowFactory
+import utopia.vault.sql.OrderDirection.Descending
+import utopia.vault.sql.{Condition, OrderDirection}
 
 object LatestModelAccess
 {
@@ -18,8 +15,8 @@ object LatestModelAccess
 	  * @tparam A Type of accessed items
 	  * @return A new access point using the specified factory and applying the specified filter condition
 	  */
-	def apply[A](factory: FromRowFactoryWithTimestamps[A], condition: Option[Condition] = None) =
-		new LatestModelAccessWrapper[A](factory, condition)
+	def apply[A](factory: FromTimelineRowFactory[A], condition: Option[Condition] = None) =
+		LatestOrEarliestModelAccess.latest(factory, condition)
 	
 	
 	// NESTED   -----------------------------
@@ -30,9 +27,10 @@ object LatestModelAccess
 	  * @param accessCondition Global condition to apply (default = None)
 	  * @tparam A Type of accessed items
 	  */
-	class LatestModelAccessWrapper[+A](override val factory: FromRowFactoryWithTimestamps[A],
+	@deprecated("Deprecated for removal. Please use LatestOrEarliestModelAccess instead", "v1.19")
+	class LatestModelAccessWrapper[+A](override val factory: FromTimelineRowFactory[A],
 	                                           override val accessCondition: Option[Condition] = None)
-		extends LatestModelAccess[A] with FilterableView[LatestModelAccessWrapper[A]]
+		extends LatestModelAccess[A]
 	{
 		override protected def self = this
 		
@@ -46,10 +44,7 @@ object LatestModelAccess
   * @author Mikko Hilpinen
   * @since 17.6.2021, v1.8
   */
-trait LatestModelAccess[+A] extends SingleRowModelAccess[A] with DistinctReadModelAccess[A, Option[A], Value]
+trait LatestModelAccess[+A] extends LatestOrEarliestModelAccess[A]
 {
-	// COMPUTED ----------------------------
-	
-	// This access point requires a timestamp-based factory
-	override def factory: FromRowFactoryWithTimestamps[A]
+	override protected def orderDirection: OrderDirection = Descending
 }